Split StorageReaderWorker once more
